feat: server socket to receive connections from SSLProxy
To intercept all messages, iptables is going to be used. After the interception, SSLProxy (github/sonartari) will be
used to decrypt the packages and send them to the extern program.

In this server socket, all the messages will be received, to be changed and sent to the socket informed by SSLProxy (read
SSLProxy documentation).