chore: add certificate to double handshake
In order to create a TLS connection with a client (firefox, curl, etc.),
we need to provide a certificate. This certificate was generated by
OpenSSL, and it's self-signed.

To allow this certificate to be accepted, it should be imported on
Firefox, or in curl, you should use the flag --insecure.